Add files via upload
Began Logistic Regression modeling. Need to really clean up the code and make it more modular.

This project has taken on a life of its own and become something other than I planned on due to the nature of the data. With the data being categorical instead of continuous different analysis methods were needed. Before moving on with additional models, the following needs to be the focus:

- code clean up
- user input options
- test files
- finish debugging logreg model